requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
groq==0.12.0
apscheduler==3.10.4
werkzeug==3.1.3
